✨ FEAT(signup) : 비밀번호 input 형식 에러 추가
- 기존의 이메일 유효성 검사 함수 이름 변경
- 기존의 이메일 유효성 검사 함수를 validations.js 모듈로 분리
- 비밀번호 input이
  - 8자 미만
  - 문자만 있는 경우
  - 숫자만 있는 경우
위의 3가지 조건을 하나라도 만족하면 에러 메시지 출력하는 이벤트 핸들러 추가
